+++ src/Middleware/Container/PrismicTemplateFactory.php
@@ -4,18 +4,18 @@ declare(strict_types=1);
 namespace Primo\Middleware\Container;
 
 use Mezzio\Template\TemplateRendererInterface;
-use Primo\Middleware\PrismicTemplateHandler;
+use Primo\Middleware\PrismicTemplate;
 use Psr\Container\ContainerInterface;
 
-final class PrismicTemplateHandlerFactory
+final class PrismicTemplateFactory
 {
-    public function __invoke(ContainerInterface $container) : PrismicTemplateHandler
+    public function __invoke(ContainerInterface $container) : PrismicTemplate
     {
         $config = $container->has('config') ? $container->get('config') : [];
         $templateAttribute = $config['primo']['templates']['templateAttribute']
-            ?? PrismicTemplateHandler::DEFAULT_TEMPLATE_ATTRIBUTE;
+            ?? PrismicTemplate::DEFAULT_TEMPLATE_ATTRIBUTE;
 
-        return new PrismicTemplateHandler(
+        return new PrismicTemplate(
             $container->get(TemplateRendererInterface::class),
             $templateAttribute
         );