+++ src/Middleware/PrismicTemplate.php
@@ -5,14 +5,14 @@ namespace Primo\Middleware;
 
 use Laminas\Diactoros\Response\HtmlResponse;
 use Mezzio\Template\TemplateRendererInterface;
-use Primo\Exception\DocumentNotFound;
 use Primo\Exception\RequestError;
 use Prismic\Document;
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;
+use Psr\Http\Server\MiddlewareInterface;
 use Psr\Http\Server\RequestHandlerInterface;
 
-final class PrismicTemplateHandler implements RequestHandlerInterface
+final class PrismicTemplate implements MiddlewareInterface
 {
     public const DEFAULT_TEMPLATE_ATTRIBUTE = 'template';
 
@@ -27,7 +27,7 @@ final class PrismicTemplateHandler implements RequestHandlerInterface
         $this->templateAttribute = $templateAttribute;
     }
 
-    public function handle(ServerRequestInterface $request) : ResponseInterface
+    public function process(ServerRequestInterface $request, RequestHandlerInterface $handler) : ResponseInterface
     {
         $template = $request->getAttribute($this->templateAttribute);
         if (! $template) {
@@ -36,7 +36,7 @@ final class PrismicTemplateHandler implements RequestHandlerInterface
 
         $document = $request->getAttribute(Document::class);
         if (! $document instanceof Document) {
-            throw DocumentNotFound::with($request);
+            return $handler->handle($request);
         }
 
         $this->renderer->addDefaultParam(